*** empty log message ***
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@3434 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/technology/technologies/Artwork.java
@@ -810,7 +810,7 @@ public class Artwork extends Technology
 				if (len == 17)
 				{
 					// the last entry specifies the outline texture
-					int outlineIndex = pat[16];
+					int outlineIndex = pat[16].intValue();
 					graphics.setOutlined(EGraphics.Outline.findOutline(outlineIndex));
 					len = 16;
 				}
